+++ cmake/rtrp.cmake
@@ -3,7 +3,7 @@ include(FetchContent)
 FetchContent_Declare(
     RTRP
     GIT_REPOSITORY https://github.com/SpaghettDev/RTRP.git
-    GIT_TAG        9309faa447abf75b11c341c9345b3d9626c20f3d
+    GIT_TAG        6a381e5d9f90ebe1640db73c8a4ba20b69577f45
     GIT_PROGRESS TRUE
 )
 message("Fetching RTRP")