+++ src/custom_layers/RLStatsAlertLayer.cpp
@@ -7,12 +7,6 @@
 #include <Geode/binding/ButtonSprite.hpp>
 #include <Geode/binding/GJLevelList.hpp>
 #include <Geode/binding/LevelListLayer.hpp>
-#include <Geode/binding/LoadingCircle.hpp>
-#include <Geode/binding/LevelCell.hpp>
-#include <Geode/binding/GJListLayer.hpp>
-#include <Geode/binding/BoomListView.hpp>
-#include <Geode/binding/TableView.hpp>
-#include <Geode/binding/CCContentLayer.hpp>
 #include <Geode/binding/TextArea.hpp>
 
 #include "RouletteManager.hpp"
@@ -24,7 +18,7 @@ RLStatsAlertLayer* RLStatsAlertLayer::create(const GameState& gs)
 {
 	auto ret = new RLStatsAlertLayer();
 
-	if (ret && ret->init(std::move(gs)))
+	if (ret && ret->init(gs))
 		ret->autorelease();
 	else
 	{